+++ src/main/java/com/oms/dto/InventoryRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class InventoryRequest {
-    private Long productId;     // Which product
-    private Long warehouseId;   // Which warehouse
+    private int productId;     // Which product
+    private int warehouseId;   // Which warehouse
     private int quantity;       // Stock to add/update
 }
